refactor: move disposition constants and functions to attachments module
