+++ src/helpers/attachments.ts
@@ -11,11 +11,7 @@ import { reduce } from 'lodash';
 
 import {
 	areContentIdsEqual,
-	DISPOSITION_ATTACHMENT,
-	DISPOSITION_INLINE,
 	extractContentIdsFromHtml,
-	isAttachmentDisposition,
-	isInlineDisposition,
 	removeAngleBrackets
 } from 'commons/content-id-utils';
 import { calcColor } from 'commons/utilities';
@@ -27,6 +23,30 @@ import {
 	SavedAttachment,
 	UnsavedAttachment
 } from 'types/index.d';
+
+/**
+ * Content disposition types for email attachments
+ */
+export const DISPOSITION_INLINE = 'inline' as const;
+export const DISPOSITION_ATTACHMENT = 'attachment' as const;
+
+/**
+ * Checks if a content disposition indicates an inline attachment.
+ *
+ * @param disposition - The content disposition value
+ * @returns True if the disposition is 'inline'
+ */
+export const isInlineDisposition = (disposition?: string): boolean =>
+	disposition === DISPOSITION_INLINE;
+
+/**
+ * Checks if a content disposition indicates a regular attachment.
+ *
+ * @param disposition - The content disposition value
+ * @returns True if the disposition is 'attachment'
+ */
+export const isAttachmentDisposition = (disposition?: string): boolean =>
+	disposition === DISPOSITION_ATTACHMENT;
 
 const FileExtensionRegex = /^.+\.([^.]+)$/;
 export const CIDURL_REGEX = '^(?:cid:)*(.+)$';
